+++ domain/src/main/java/vzijden/workout/domain/model/PlannedSet.kt
@@ -4,13 +4,13 @@ package vzijden.workout.domain.model
 data class PlannedSet(
 
     var reps: Int,
-    var registrationId: Int
+    var registrationId: Long
 ) {
 
-  var id: Int = 0
+  var id: Long = 0
     private set
 
-  constructor(reps: Int, registrationId: Int, id: Int) : this(reps, registrationId) {
+  constructor(reps: Int, registrationId: Long, id: Long) : this(reps, registrationId) {
     this.id = id
   }
 }
